Refactoring the `SingleFileMetastore` to prepare for #839.
- The lock is now per index.
- Factored together the mutation and the view logic to improve
  proofreadability.
- Discarding entry in the cache after an error.
